+++ packages/utilities/src/index.ts
@@ -4,3 +4,4 @@ export * from "./context-utils";
 export * from "./assertion";
 export * from "./error-utils";
 export * from "./wave-utils";
+export * from "./service-utils";