+++ services/assets/serverless.ts
@@ -2,9 +2,9 @@ import type { AWS } from "@serverless/typescript";
 import { functions } from "./functions/index";
 
 const serverlessConfiguration: AWS = {
-  service: "deploy",
+  service: "assets",
   frameworkVersion: "3.34.0",
-  plugins: ["serverless-esbuild", "serverless-offline", "serverless-iam-roles-per-function"],
+  plugins: ["serverless-esbuild", "serverless-offline"],
   provider: {
     name: "aws",
     runtime: "nodejs16.x",
@@ -16,6 +16,7 @@ const serverlessConfiguration: AWS = {
     },
     deploymentMethod: "direct",
     apiGateway: {
+      binaryMediaTypes: ["*/*"],
       restApiId: {
         "Fn::ImportValue": "ApiRestId",
       },
@@ -30,6 +31,7 @@ const serverlessConfiguration: AWS = {
   functions: { ...functions },
   package: { individually: true },
   custom: {
+    domain: "api.vastly.is",
     "serverless-offline": {
       httpPort: 4000,
       useChildProcesses: true,
@@ -47,10 +49,10 @@ const serverlessConfiguration: AWS = {
   },
   resources: {
     Resources: {
-      deployAssumeRole: {
+      assetServiceRole: {
         Type: "AWS::IAM::Role",
         Properties: {
-          RoleName: "DeployAssumeRole",
+          RoleName: "AssetServiceRole",
           AssumeRolePolicyDocument: {
             Version: "2012-10-17",
             Statement: [
@@ -65,14 +67,14 @@ const serverlessConfiguration: AWS = {
           },
           Policies: [
             {
-              PolicyName: "SSTDeployPolicy",
+              PolicyName: "AssetServicePolicy",
               PolicyDocument: {
                 Version: "2012-10-17",
                 Statement: [
                   {
                     Effect: "Allow",
-                    Action: ["cloudformation:*", "s3:*", "iam:*", "lambda:*", "ecr:*", "ssm:*"],
-                    Resource: ["*"],
+                    Action: "s3:*",
+                    Resource: "*",
                   },
                 ],
               },
@@ -82,9 +84,9 @@ const serverlessConfiguration: AWS = {
       },
     },
     Outputs: {
-      DeployAssumeRoleOutput: {
-        Value: { Ref: "deployAssumeRole" },
-        Export: { Name: "DeployAssumeRoleArn-${self:provider.stage}" },
+      AssetServiceRoleOutput: {
+        Value: { "Fn::GetAtt": ["assetServiceRole", "Arn"] },
+        Export: { Name: "AssetServiceRoleArn" },
       },
     },
   },