+++ services/environment/CHANGELOG.md
@@ -1,5 +1,12 @@
 # environment
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @vastly/utils@0.4.1
+
 ## 0.5.1
 
 ### Patch Changes